Change math to 2D/3D variant, change doubles->floats

+++ src/core/aabb.h
@@ -7,7 +7,7 @@ struct AABB {
   // Bounding box with Inifinite bounds
   AABB() : min(Vec3(INFINITY)), max(Vec3(-INFINITY)) {};
   // Bounding box from (-v, -v, -v) to (v, v, v)
-  AABB(double v) : min(Vec3(-v)), max(Vec3(v)) {};
+  AABB(float v) : min(Vec3(-v)), max(Vec3(v)) {};
   // Create Bounds with a single point
   AABB(Vec3 v) : min(v), max(v) {};
   // Bounding box with `min` and `max` bounds
@@ -15,8 +15,8 @@ struct AABB {
   
   void addPoint(const Vec3& p);
   Vec3 offset(const Vec3& p) const;
-  bool hit(Ray &ray, double &t1, double &t2, const Vec3& invD);
-  bool hit(Ray &ray, double &t1, double &t2);
+  bool hit(Ray &ray, float &t1, float &t2, const Vec3& invD);
+  bool hit(Ray &ray, float &t1, float &t2);
   bool hit(Ray &ray);
   Vec3 min;
   Vec3 max;
@@ -28,5 +28,5 @@ AABB combine(const AABB& a, const AABB& b);
 AABB transformed(const Matrix4& T, const AABB& b);
 bool overlap(const AABB& a, const AABB& b);
 Vec3 centroid(const AABB& a);
-double area(const AABB& a);
+float area(const AABB& a);
