+++ src/integrators/sppm.cc
@@ -39,7 +39,7 @@ void SPPM::addVisiblePoint(const Ray& r, Scene *scene, RNG& rng, int pixIdx) {
 void SPPM::sendPhoton(Scene *scene, RNG &rng) {
   if (scene->lights.size() == 0) return; 
 
-  double pdf;
+  float pdf;
 
   // Pick a light source, and pick a point on it
   Object *light = scene->lights[ rng.randint() % scene->lights.size() ];
