+++ src/objects/displaced.h
@@ -10,7 +10,7 @@ struct DisplacedSphere : ParametricSurface {
   DisplacedSphere(
     Texture *displacementMap, 
     BSDF *mat, 
-    double scale=.2, 
+    float scale=.2, 
     int resolution = 100, 
     bool interpNormals=true
   ) : ParametricSurface(mat, resolution, resolution, 0, TAU, EPS, PI-EPS),
@@ -19,15 +19,15 @@ struct DisplacedSphere : ParametricSurface {
     interpolateNormals = interpNormals;
   };
 
-  Vec3 P(double a, double b) {
-    double u = map(a, aMin, aMax, 0, 1);
-    double v = map(b, bMin, bMax, 0, 1);
-    double offset = luminance(displacementMap->get(u, v)) * scale;
+  Vec3 P(float a, float b) {
+    float u = map(a, aMin, aMax, 0, 1);
+    float v = map(b, bMin, bMax, 0, 1);
+    float offset = luminance(displacementMap->get(u, v)) * scale;
     
     Vec3 pt = Vec3(cos(a) * sin(b), sin(a) * sin(b), cos(b));
     return pt * (1+offset);
   }
-  double scale;
+  float scale;
   Texture *displacementMap;
 };
 
@@ -35,7 +35,7 @@ struct DisplacedPlane : ParametricSurface {
   DisplacedPlane(
     Texture *displacementMap, 
     BSDF *mat, 
-    double scale=.2,
+    float scale=.2,
     int resolution = 100, 
     bool interpNormals=true
   ) : ParametricSurface(mat, resolution, resolution, -1, 1, -1, 1),
@@ -44,12 +44,12 @@ struct DisplacedPlane : ParametricSurface {
     interpolateNormals = interpNormals;
   };
 
-  Vec3 P(double a, double b) {
-    double u = map(a, aMin, aMax, 0, 1);
-    double v = map(b, bMin, bMax, 0, 1);
-    double offset = luminance(displacementMap->get(u, v)) * scale;
+  Vec3 P(float a, float b) {
+    float u = map(a, aMin, aMax, 0, 1);
+    float v = map(b, bMin, bMax, 0, 1);
+    float offset = luminance(displacementMap->get(u, v)) * scale;
     return Vec3(a, b, offset);
   }
-  double scale;
+  float scale;
   Texture *displacementMap;
 };
