+++ src/objects/plane.h
@@ -7,6 +7,6 @@ Plane(BSDF *mat) : Object(mat) {
     bounds = AABB(Vec3(-1, -1, 0), Vec3(1, 1, 0));
   };
   bool hit(Ray& r, HitRec &rec);
-  Vec3 sample(double *pdf, RNG& rng);
+  Vec3 sample(float *pdf, RNG& rng);
 };
